(qtcreator-cdbext) Add Qt Creator CDB Extension
This adds the CDB debugger extension for Qt Creator,
which is provided in a separate .7z file on Qt's
download page.

Since it is an extension for Qt Creator, this
package depends on the 'qtcreator' package.

The packaging files were created based on the ones
for the 'qtcreator' package, i.e. those for the
existing 'qtcreator' package were initially
copied, then adapted for this new package as
needed.

Fixes: #216